Move wfss_multispec_to_source to exp_to_source from wfss_multispec module as requested by Ned Molter, Melanie Clarke, and Tyler Pauly
